requirements/requirements.txt
Django
django-guardian
django-registration-redux
django-crispy-forms
Jinja2
psycopg2
pytz
envparse
